requirements.txt
Django==4.2
djangorestframework==3.14.0
djoser==2.1.0
environs==9.5.0
mysqlclient==2.1.1